requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
lxml_html_clean==0.1.1
selenium==4.18.1
webdriver-manager==4.0.1
kaggle==1.6.6
python-dotenv==1.0.0
html5lib==1.1
newspaper3k==0.2.8
google-api-python-client==2.108.0
scikit-learn==1.3.2
openai==1.12.0
